test: split globalAgent TLS tests into smaller files for ASAN performance
Split globalAgent-tls-options.test.ts (19 tests) into 3 smaller files:
- globalAgent-https-request.test.ts (4 tests)
- globalAgent-fetch.test.ts (7 tests)
- globalAgent-undici.test.ts (8 tests)

Changes:
- Use bunRun (spawnSync) instead of async Bun.spawn for simpler code
- Add describe.concurrent for parallel test execution within files
- Reduces ASAN build timeout issues by splitting subprocess spawns
